Make an opt-out path for the BINARY format
Make it possible for BasicTracer users to avoid even an *import* of the
BinaryPropagator, as the latter can throw protobuf-related exceptions at
initialization time.
